requirements.txt
streamlit==1.32.0
ultralytics
langchain==0.1.12
langchain-groq==0.1.0
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4 